OBPIH-6850 Move pagination to a separate hook
