+++ src/js/components/DataTable/v2/DataTable.jsx
@@ -12,33 +12,21 @@ import 'components/DataTable/DataTable.scss';
 const DataTable = ({
   columns,
   data,
-  pageSize,
   footerComponent,
   loading,
   loadingMessage,
   emptyTableMessage,
-  setOffset,
-  setPageSize,
   totalCount,
   filterParams,
+  paginationProps,
 }) => {
   const {
     defaultEmptyTableMessage,
     defaultLoadingTableMessage,
-    canPrevious,
-    canNext,
-    pageSizeSelectOptions,
-    onPageSizeChange,
-    onPageChange,
     table,
-    pages,
-    pagination,
   } = useDataTable({
-    pageSize,
     columns,
     data,
-    setOffset,
-    setPageSize,
     totalCount,
     filterParams,
   });
@@ -61,14 +49,8 @@ const DataTable = ({
           />
           <DataTableFooter
             footerComponent={footerComponent}
-            pagination={pagination}
-            onPageChange={onPageChange}
-            onPageSizeChange={onPageSizeChange}
-            canPrevious={canPrevious}
-            canNext={canNext}
-            pages={pages}
-            pageSizeSelectOptions={pageSizeSelectOptions}
             totalData={totalCount}
+            {...paginationProps}
           />
         </div>
       </div>
@@ -85,7 +67,6 @@ DataTable.propTypes = {
   data: PropTypes.arrayOf(
     PropTypes.shape({}),
   ).isRequired,
-  pageSize: PropTypes.number,
   footerComponent: PropTypes.func,
   loading: PropTypes.bool,
   loadingMessage: PropTypes.shape({
@@ -108,7 +89,6 @@ DataTable.defaultProps = {
   loading: false,
   loadingMessage: null,
   emptyTableMessage: null,
-  pageSize: 5,
   setOffset: () => {},
   totalCount: 0,
   setPageSize: () => {},
